Backend/project/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
asyncio==3.4.3
blinker==1.7.0
boto3==1.34.35
botocore==1.34.35
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fiscalyear==0.4.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
openai==1.11.1
packaging==23.2
parso==0.8.3
pathlib==1.0.1
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pymssql==2.2.11
pytesseract==0.3.10
python-dateutil==2.8.2
razorpay==1.4.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
sniffio==1.3.0
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==3.0.1
zipp==3.17.0
